♻️ [refactor] #5 CH2 키오스크 과제 메뉴 아이템을 한 번에 추가할 수 있도록 List.of()를 활용하여 초기화
