+++ src/main/java/org/KioskProject/mandatory/level45/Menu.java
@@ -1,19 +1,18 @@
 package org.KioskProject.mandatory.level45;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class Menu {
     // 카테고리 이름 필드 추가
-    private String categoryName;
+    private final String categoryName;
 
     // MenuItem 클래스를 List로 관리
-    private List<MenuItem> menuItems;
+    private final List<MenuItem> menuItems;
 
     // 생성자
-    public Menu(String categoryName) {
+    public Menu(String categoryName, List<MenuItem> menuItems) {
         this.categoryName = categoryName;
-        this.menuItems = new ArrayList<>();
+        this.menuItems  = menuItems;
     }
 
     // List에 포함된 MenuItem을 순차적으로 보여주는 함수
